Lmdb poc (#166)
* adding first version of lmdb

* cleanup

* adding expiry

* adding lmdb to scopez server and bb test

* adding more bb tests

* cleanup

* adding lmdb

* cleanup

+++ src/db/kycb_db.cc
@@ -358,7 +358,7 @@ int32_t kycb_db::print_all_keys(void)
 //: \return:  TODO
 //: \param:   TODO
 //: ----------------------------------------------------------------------------
-int32_t kycb_db::set_opt(uint32_t a_opt, const void *a_buf, uint32_t a_len)
+int32_t kycb_db::set_opt(uint32_t a_opt, const void *a_buf, uint64_t a_len)
 {
         switch(a_opt)
         {
